backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
hdbscan==0.8.33
scikit-learn==1.3.2
numpy==1.26.2
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0